bug fix: add missing parentheses

+++ artemis/modules/nuclei.py
@@ -460,8 +460,8 @@ class Nuclei(ArtemisBase):
             )
         )
 
-        if capped := Config.Modules.Nuclei.NUCLEI_MAX_SECONDS_PER_REQUEST_ON_RETRY > 0:
-            milliseconds_per_request_retry = min(milliseconds_per_request_retry, 1000 * capped)
+        if (capped := Config.Modules.Nuclei.NUCLEI_MAX_SECONDS_PER_REQUEST_ON_RETRY) > 0:
+            milliseconds_per_request_retry = min(milliseconds_per_request_retry, int(1000 * capped))
 
         milliseconds_per_request_candidates = [milliseconds_per_request_initial, milliseconds_per_request_retry]
